Fix DOM nesting error
