+++ frontend/src/App.tsx
@@ -5,6 +5,7 @@ import HomePage from './pages/HomePage';
 import FAQPage from './pages/FAQPage';
 import LandlordPage from './pages/LandlordPage';
 import { createMuiTheme, ThemeProvider } from '@material-ui/core';
+import NavBar, { NavbarButton } from './components/utils/NavBar';
 
 const theme = createMuiTheme({
   palette: {
@@ -20,10 +21,23 @@ const theme = createMuiTheme({
   },
 });
 
+const faq: NavbarButton = {
+  label: 'FAQ',
+  href: '/faq',
+};
+
+const review: NavbarButton = {
+  label: 'Reviews',
+  href: '/landlord/1',
+};
+
+const headersData = [faq, review];
+
 const App = (): ReactElement => {
   return (
     <ThemeProvider theme={theme}>
       <Router>
+        <NavBar headersData={headersData} />
         <Switch>
           <Route exact path="/" component={HomePage} />
           <Route exact path="/faq" component={FAQPage} />
