+++ frontend/src/pages/LandlordPage.module.scss
@@ -1,0 +1,3 @@
+.OuterContainer {
+  margin-top: 1em;
+}
